+++ InstantMock/Arguments/Captor/ArgumentCaptor.swift
@@ -18,6 +18,9 @@ public class ArgumentCaptor<T> {
         return ArgumentCaptorValuesImpl<T>(self.arg)
     }()
 
+    /** Main initializer */
+    public init() {}
+
 }